mediscript_ai/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
alembic
psycopg2-binary
pydantic
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
Pillow
rapidfuzz
reportlab
pydantic-settings
pytest
google-generativeai
httpx
